requirements.txt
opencv-python
numpy
pupil-apriltags
svgwrite
reportlab
moms-apriltag
pillow
